XRDDEV-1639 Move key column sorting to it's own file
